Refactored resnet.py to split into two files
