+++ keras_cv/models/__init__.py
@@ -15,9 +15,9 @@
 from keras_cv.models.densenet import DenseNet121
 from keras_cv.models.densenet import DenseNet169
 from keras_cv.models.densenet import DenseNet201
-from keras_cv.models.resnet import ResNet50
-from keras_cv.models.resnet import ResNet50V2
-from keras_cv.models.resnet import ResNet101
-from keras_cv.models.resnet import ResNet101V2
-from keras_cv.models.resnet import ResNet152
-from keras_cv.models.resnet import ResNet152V2
+from keras_cv.models.resnet_v1 import ResNet50
+from keras_cv.models.resnet_v1 import ResNet101
+from keras_cv.models.resnet_v1 import ResNet152
+from keras_cv.models.resnet_v2 import ResNet50V2
+from keras_cv.models.resnet_v2 import ResNet101V2
+from keras_cv.models.resnet_v2 import ResNet152V2
